# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuardianEye API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface Profile {
  id: string;
  /** @nullable */
  display_name?: string | null;
  /** @nullable */
  avatar_url?: string | null;
  /** @nullable */
  pin_hash?: string | null;
  subscription_tier: string;
  theme: string;
  created_at: string;
  updated_at: string;
}

export interface ProfileUpdate {
  display_name?: string;
}

export interface PinInput {
  pin_hash: string;
}

export interface PinVerifyResult {
  valid: boolean;
}

export interface Child {
  id: string;
  name: string;
  /** @nullable */
  avatar_url?: string | null;
  /** @nullable */
  birth_year?: number | null;
  /** @nullable */
  color?: string | null;
  parent_id: string;
  created_at: string;
}

export interface ChildInput {
  name: string;
  birth_year?: number;
  color?: string;
}

export interface Alert {
  id: string;
  /** @nullable */
  child_id?: string | null;
  parent_id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  severity: string;
  read: boolean;
  created_at: string;
}

export interface AppLimit {
  id: string;
  child_id: string;
  parent_id: string;
  app_name: string;
  /** @nullable */
  package_id?: string | null;
  daily_minutes: number;
  blocked: boolean;
  created_at: string;
}

export interface AppLimitInput {
  app_name: string;
  package_id?: string;
  daily_minutes?: number;
  blocked?: boolean;
}

export interface AppLimitUpdate {
  app_name?: string;
  daily_minutes?: number;
  blocked?: boolean;
}

export interface Device {
  id: string;
  child_id: string;
  parent_id: string;
  /** @nullable */
  device_name?: string | null;
  /** @nullable */
  platform?: string | null;
  status: string;
  /** @nullable */
  pairing_code?: string | null;
  /** @nullable */
  paired_at?: string | null;
  /** @nullable */
  last_seen?: string | null;
  created_at: string;
}

export interface DeviceInput {
  device_name: string;
  platform?: string;
}

export interface WebBlocklistEntry {
  id: string;
  child_id: string;
  parent_id: string;
  domain: string;
  /** @nullable */
  category?: string | null;
  created_at: string;
}

export interface WebBlocklistInput {
  domain: string;
  category?: string;
}

export interface ChildOverview {
  child: Child;
  recent_alerts: Alert[];
  unread_alerts_count: number;
  active_devices_count: number;
  app_limits_count: number;
  blocked_sites_count: number;
}