# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuardianEye API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Alert {
  id: string;
  /** @nullable */
  child_id?: string | null;
  parent_id: string;
  title: string;
  /** @nullable */
  description?: string | null;
  severity: string;
  read: boolean;
  created_at: string;
}