# lib/api-zod/src/generated/types/appLimit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuardianEye API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AppLimit {
  id: string;
  child_id: string;
  parent_id: string;
  app_name: string;
  /** @nullable */
  package_id?: string | null;
  daily_minutes: number;
  blocked: boolean;
  created_at: string;
}