# lib/api-zod/src/generated/types/appLimitInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuardianEye API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AppLimitInput {
  app_name: string;
  package_id?: string;
  daily_minutes?: number;
  blocked?: boolean;
}