# lib/api-zod/src/generated/types/child.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuardianEye API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Child {
  id: string;
  name: string;
  /** @nullable */
  avatar_url?: string | null;
  /** @nullable */
  birth_year?: number | null;
  /** @nullable */
  color?: string | null;
  parent_id: string;
  created_at: string;
}