# lib/api-zod/src/generated/types/childOverview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuardianEye API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Alert } from "./alert";
import type { Child } from "./child";

export interface ChildOverview {
  child: Child;
  recent_alerts: Alert[];
  unread_alerts_count: number;
  active_devices_count: number;
  app_limits_count: number;
  blocked_sites_count: number;
}