# lib/api-zod/src/generated/types/device.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuardianEye API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Device {
  id: string;
  child_id: string;
  parent_id: string;
  /** @nullable */
  device_name?: string | null;
  /** @nullable */
  platform?: string | null;
  status: string;
  /** @nullable */
  pairing_code?: string | null;
  /** @nullable */
  paired_at?: string | null;
  /** @nullable */
  last_seen?: string | null;
  created_at: string;
}