# lib/api-zod/src/generated/types/deviceInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuardianEye API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DeviceInput {
  device_name: string;
  platform?: string;
}