# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuardianEye API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: string;
  /** @nullable */
  display_name?: string | null;
  /** @nullable */
  avatar_url?: string | null;
  /** @nullable */
  pin_hash?: string | null;
  subscription_tier: string;
  theme: string;
  created_at: string;
  updated_at: string;
}