# lib/api-zod/src/generated/types/webBlocklistEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GuardianEye API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WebBlocklistEntry {
  id: string;
  child_id: string;
  parent_id: string;
  domain: string;
  /** @nullable */
  category?: string | null;
  created_at: string;
}